v3.9.9.58: Enable bill suffix lookup, decimal stocktake inputs, non-consumables light POS theme, per-waiter per-shift bill isolation & horizontal table scroll tools

+++ famous_gates_app/lib/features/branch_accountant/presentation/bar_stocktake_review_screen.dart
@@ -128,7 +128,7 @@ class _BarStocktakeReviewScreenState
     final sdds = -additions;
     final physical = _nullableNum(
       r['physical_quantity'] ?? r['counted_quantity'] ?? r['actual_quantity'],
-    )?.toInt();
+    )?.toDouble();
     final name = '${r['item_name'] ?? r['name'] ?? 'Item'}';
     return StockTakeItem(
       id: '${r['item_id'] ?? r['id']}',
